[fixup]: Fixed invalid render surface being used on minimized.

+++ Animation/AnimationChannel.h
@@ -4,27 +4,27 @@
 template<typename T>
 struct AnimationKeyframe {
     T value;
-    float time;
+    double time;
 };
 
 template<typename T>
 struct KeyFrameInterpolation {
-    static T Interpolate(T start, T, float) {
+    static T Interpolate(T start, T, double) {
         return start;
     }
 };
 
 template<>
 struct KeyFrameInterpolation<glm::vec3> {
-    static  glm::vec3 Interpolate(glm::vec3 start, glm::vec3 end, float t) {
+    static  glm::vec3 Interpolate(glm::vec3 start, glm::vec3 end, double t) {
         return glm::mix(start, end, t);
     }
 };
 
 template<>
 struct KeyFrameInterpolation<glm::quat> {
-    static glm::quat Interpolate(glm::quat start, glm::quat end, float t) {
-        return glm::slerp(start, end, t);
+    static glm::quat Interpolate(glm::quat start, glm::quat end, double t) {
+        return glm::slerp(start, end, (float)t);
     }
 };
 
@@ -33,7 +33,7 @@ struct KeyFrameInterpolation<glm::quat> {
 class AnimationChannel {
 public:
     template<typename T, typename keyframe_type = AnimationKeyframe<T>>
-    static std::pair<keyframe_type, keyframe_type> GetSurroundingKeyframes(float time, const std::vector<keyframe_type>& keyframes) {
+    static std::pair<keyframe_type, keyframe_type> GetSurroundingKeyframes(double time, const std::vector<keyframe_type>& keyframes) {
         if(keyframes.empty()) {
             return std::make_pair(keyframe_type{}, keyframe_type{});
         }
@@ -59,16 +59,16 @@ public:
     }
 
     template<typename T, typename keyframe_type = AnimationKeyframe<T>>
-    static T GetInterpolatedValue(float time, const std::vector<keyframe_type>& keyframes) {
+    static T GetInterpolatedValue(double time, const std::vector<keyframe_type>& keyframes) {
         auto [start, end] = GetSurroundingKeyframes<T>(time, keyframes);
-        float t = (time - start.time) / (end.time - start.time);
+        double t = (time - start.time) / (end.time - start.time);
         if(end.time - start.time == 0) {
             t = 1.0f;
         }
         return KeyFrameInterpolation<T>::Interpolate(start.value, end.value, t);
     }
 
-    glm::mat4 GetInterpolatedTransform(float time) const {
+    glm::mat4 GetInterpolatedTransform(double time) const {
         auto transform = glm::mat4(1.0f);
         transform = glm::toMat4(GetInterpolatedValue<glm::quat>(time, rotation_keyframes)) * transform;
         //transform = glm::scale(glm::mat4(1.0f), GetInterpolatedValue<glm::vec3>(time, scale_keyframes)) * transform;
@@ -78,15 +78,15 @@ public:
 
     const std::string& GetName() const { return name; }
 
-    float GetDuration() const { return duration; }
-    float GetTicksPerSecond() const { return ticks_per_second; }
+    double GetDuration() const { return duration; }
+    double GetTicksPerSecond() const { return ticks_per_second; }
 
 private:
     friend class AnimationManager;
     std::vector<AnimationKeyframe<glm::vec3>> position_keyframes;
     std::vector<AnimationKeyframe<glm::vec3>> scale_keyframes;
     std::vector<AnimationKeyframe<glm::quat>> rotation_keyframes;
-    float duration;
-    float ticks_per_second;
+    double duration;
+    double ticks_per_second;
     std::string name;
 };
